libs/Ncxx/Nc3File - changing use of int64_t to long long

+++ codebase/libs/Ncxx/src/include/Ncxx/Nc3File.hh
@@ -117,7 +117,7 @@ public:
   Nc3Bool add_att( Nc3Token attname, short );
   Nc3Bool add_att( Nc3Token attname, long );
   Nc3Bool add_att( Nc3Token attname, int );
-  Nc3Bool add_att( Nc3Token attname, int64_t );
+  Nc3Bool add_att( Nc3Token attname, long long );
   Nc3Bool add_att( Nc3Token attname, float );
   Nc3Bool add_att( Nc3Token attname, double );
   Nc3Bool add_att( Nc3Token attname, const char*);       // string attribute
@@ -126,7 +126,7 @@ public:
   Nc3Bool add_att( Nc3Token attname, int, const short* );
   Nc3Bool add_att( Nc3Token attname, int, const long* );
   Nc3Bool add_att( Nc3Token attname, int, const int* );
-  Nc3Bool add_att( Nc3Token attname, int, const int64_t* );
+  Nc3Bool add_att( Nc3Token attname, int, const long long* );
   Nc3Bool add_att( Nc3Token attname, int, const float* );
   Nc3Bool add_att( Nc3Token attname, int, const double* );
 
@@ -330,7 +330,7 @@ public:
   Nc3Bool add_att( Nc3Token, short );
   Nc3Bool add_att( Nc3Token, int );
   Nc3Bool add_att( Nc3Token, long );
-  Nc3Bool add_att( Nc3Token, int64_t);
+  Nc3Bool add_att( Nc3Token, long long);
   Nc3Bool add_att( Nc3Token, float );
   Nc3Bool add_att( Nc3Token, double );
   Nc3Bool add_att( Nc3Token, const char* );      // string attribute
@@ -338,7 +338,7 @@ public:
   Nc3Bool add_att( Nc3Token, int, const ncbyte* );
   Nc3Bool add_att( Nc3Token, int, const short* );
   Nc3Bool add_att( Nc3Token, int, const int* );
-  Nc3Bool add_att( Nc3Token, int, const int64_t* );
+  Nc3Bool add_att( Nc3Token, int, const long long* );
   Nc3Bool add_att( Nc3Token, int, const long* );
   Nc3Bool add_att( Nc3Token, int, const float* );
   Nc3Bool add_att( Nc3Token, int, const double* );
